[WIP] Assembly tests run; need to connect trace port
